requirements.txt
boto3==1.18.21
pandas==1.3.4
psycopg2==2.9.2
python-dotenv==0.19.2
